requirements.txt
absl-py==1.4.0
attrs==22.2.0
BaseDT==0.1.3
BaseML==0.1.5
cffi==1.17.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
flatbuffers==24.3.25
fonttools==4.39.3
imageio==2.35.1
importlib-resources==5.12.0
joblib==1.4.2
kiwisolver==1.4.4
lazy-loader==0.4
matplotlib==3.7.1
mediapipe==0.10.0
networkx==3.1
numpy==1.24.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==23.0
pandas==2.0.3
Pillow==9.4.0
protobuf==3.20.3
pycparser==2.22
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.2
PyWavelets==1.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sounddevice==0.5.1
threadpoolctl==3.5.0
tifffile==2023.7.10
tqdm==4.65.0
tzdata==2024.2
yellowbrick==1.5
zipp==3.15.0
